Rename graphjs.py to graphjs and add `--with-types` argument
Renames `graphjs.py` to graphjs so that it looks more like a
binary file. Also, configure `graphjs` to be able to be
executed from any directory just by adding the project root
to the PATH environment variable.

Adds `--with-types` argument to run graphjs with type
reconstruction on without having to use `-e` that
generates symbolic tests as well. Now running `-e`
implies the argument `--with-types`.